# internal/order/ports/openapi_api.gen.go
// Package ports provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.16.3 DO NOT EDIT.
package ports

import (
	"fmt"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (POST /customer/{customer_id}/orders)
	PostCustomerCustomerIdOrders(c *gin.Context, customerId string)

	// (GET /customer/{customer_id}/orders/{orders_id})
	GetCustomerCustomerIdOrdersOrdersId(c *gin.Context, customerId string, ordersId string)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// PostCustomerCustomerIdOrders operation middleware
func (siw *ServerInterfaceWrapper) PostCustomerCustomerIdOrders(c *gin.Context) {

	var err error

	// ------------- Path parameter "customer_id" -------------
	var customerId string

	err = runtime.BindStyledParameter("simple", false, "customer_id", c.Param("customer_id"), &customerId)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter customer_id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostCustomerCustomerIdOrders(c, customerId)
}

// GetCustomerCustomerIdOrdersOrdersId operation middleware
func (siw *ServerInterfaceWrapper) GetCustomerCustomerIdOrdersOrdersId(c *gin.Context) {

	var err error

	// ------------- Path parameter "customer_id" -------------
	var customerId string

	err = runtime.BindStyledParameter("simple", false, "customer_id", c.Param("customer_id"), &customerId)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter customer_id: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Path parameter "orders_id" -------------
	var ordersId string

	err = runtime.BindStyledParameter("simple", false, "orders_id", c.Param("orders_id"), &ordersId)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter orders_id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetCustomerCustomerIdOrdersOrdersId(c, customerId, ordersId)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.POST(options.BaseURL+"/customer/:customer_id/orders", wrapper.PostCustomerCustomerIdOrders)
	router.GET(options.BaseURL+"/customer/:customer_id/orders/:orders_id", wrapper.GetCustomerCustomerIdOrdersOrdersId)
}